requirements.txt
chromadb==1.3.4
sentence-transformers==2.6.0
openai>=1.0.0
tavily-python>=0.3.2
python-dotenv>=1.0.1
